requirements.txt
# Lista de dependências do Python. Instale com:
# python -m pip install -r requirements.txt
fastapi>=0.111,<0.116
uvicorn[standard]>=0.30,<0.31
sqlalchemy>=2.0.31,<2.1
psycopg2-binary>=2.9,<3
pydantic>=2.8,<3
python-dotenv>=1.0,<2
passlib[bcrypt]==1.7.4
PyJWT>=2.9,<3
authlib>=1.3,<2
httpx>=0.27,<0.28
fastapi-limiter==0.1.6
redis>=5.0,<6
APScheduler>=3.10,<3.12
aiosmtplib>=2.0,<3
feedparser>=6.0,<7
email-validator
fastapi
uvicorn[standard]
httpx
pydantic[email]
